+++ src/common/index.ts
@@ -6,6 +6,7 @@ import getColorConfiguration from "./utils/getColorConfiguration";
 import { getScrollParent } from "./utils/getScrollParent";
 import { getGlobalVar, setGlobalVar } from "./utils/globalVars";
 import { openInNewTab } from "./utils/openInNewTab";
+import { reduceToText } from "./utils/reduceToText"
 export type { IntentTypes as IntentBaseTypes } from "./Intent";
 
 export const utils = {
@@ -19,4 +20,5 @@ export const utils = {
     getScrollParent,
     getEnabledColorsFromPalette,
     textToColorHash,
+    reduceToText
 };