+++ CHANGELOG.md
@@ -113,6 +113,8 @@ This is a major release, and it might be not compatible with your current usage 
     -   `intent` display was aligned with other components, `intent="primary"` is now `intent="accent"`, in most cases it may be better to use `affirmative={true}` or `elevated={true}` instead of primary/accent intent
 -   `<Spinner />`
     -   `intent` display was aligned with other components, `intent="primary"` is now `intent="accent"`, in most cases it may be better to use `elevated={true}` instead of using intent
+-  `<TextReducer>`
+    -  extracted utility for shrinking content to plain text. The utility, `reduceToText`, can be used independently from the component.
 
 ### Deprecated